more generics
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@4305 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/kelondro/kelondroCollectionIndex.java
@@ -386,7 +386,7 @@ public class kelondroCollectionIndex {
         while (i.hasNext()) {
             entry = i.next();
             actionList = entry.getValue();
-            partitionNumber = entry.getKey();
+            partitionNumber = entry.getKey().intValue();
             array = getArray(partitionNumber, serialNumber, index.row().objectOrder, chunkSize);
             j = actionList.iterator();
             while (j.hasNext()) {
